fix: Somes changes following the last review steps
